backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
ibm_db==3.2.8
ibm_db_sa==0.4.4
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.3.0
langchain-openai==1.1.14
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
